Modified plot_rd.m to read from mysql database
It works by getting all necessary data from the database, using the
problem name and batch size, then using the system to print the output
to a file in /tmp/, and loading the file to access the data.

The use of an intermediary file is not the most elegant solution, but it
is a quick solution that doesn't require correctling setting up the
MySQL database connector. The extra time required to write to and read
from the file is insignificant for the amount of data that will be used
in these plots.